+++ examples/data-sources/iosxr_l2vpn_bridge_group/data-source.tf
@@ -1,0 +1,3 @@
+data "iosxr_l2vpn_bridge_group" "example" {
+  group_name = "BG123"
+}