+++ examples/resources/iosxr_l2vpn_bridge_group/resource.tf
@@ -1,0 +1,3 @@
+resource "iosxr_l2vpn_bridge_group" "example" {
+  group_name = "BG123"
+}